requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-OAuthlib==0.9.1
Flask-RESTful==0.3.2
oauthlib==0.7.2
pydash==2.4.2
Werkzeug==0.10.4
factory_boy==2.4.1
factoryboy-gaendb==0.1.5
fake-factory==0.5.1
Unidecode==0.4.17
geojson==1.0.9
nose==1.3.7
NoseGAE==0.5.8
simplejson==3.8.2
Pillow==3.2.0
GoogleAppEngineCloudStorageClient==1.9.22.1
